requirements.txt
ddt               
playwright
rich
# tomli
python-dotenv